# openapi.yaml
openapi: 3.0.3
info:
  title: Hypernova Banking API
  version: 1.0.0
  license:
    name: Proprietary technical assessment implementation
    url: https://github.com/0xStyx507/hypernova-banking
  description: |
    Contrato HTTP de Hypernova Banking.

    Incluye identidad, sesiones, auditoría, cuentas USD y movimientos
    financieros respaldados por TigerBeetle. Los importes son unidades menores
    enteras serializadas como cadenas para evitar redondeos de punto flotante.
servers:
  - url: http://127.0.0.1:8080
    description: Desarrollo local
# Public health and authentication operations opt in to security below; an
# empty root requirement makes the security posture explicit for every path.
security: []
...
paths:
...
  /api/v1/accounts:
    post:
      operationId: createAccount
      summary: Crea otra cuenta USD para el usuario autenticado
      description: Cada cuenta pertenece al usuario autenticado y puede seleccionarse como origen de una operación. La identidad recibe una cuenta inicial durante el registro.
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Cuenta provisionada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Solicitud inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    get:
      operationId: listAccounts
      summary: Lista las cuentas propias
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Cuentas del usuario
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Account' }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/accounts/{account_id}:
    get:
      operationId: getAccount
      summary: Obtiene una cuenta propia
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountID'
      responses:
        '200':
          description: Cuenta
          content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Cuenta inexistente o ajena
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    delete:
      operationId: closeAccount
      summary: Cierra una cuenta propia con saldo cero
      description: Cierra lógicamente la cuenta y conserva sus referencias de auditoría. El saldo disponible y los movimientos pendientes deben ser cero.
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountID'
      responses:
        '204': { description: Cuenta cerrada }
        '401':
          description: Sesión inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Cuenta inexistente o ajena
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '409':
          description: La cuenta conserva saldo o movimientos pendientes
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    patch:
      operationId: renameAccount
      summary: Cambia el nombre visible de una cuenta propia
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [display_name]
              properties:
                display_name: { type: string, minLength: 2, maxLength: 48 }
      responses:
        '200':
          description: Cuenta actualizada
          content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } }
        '400':
          description: Nombre inválido
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Sesión inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Cuenta inexistente o ajena
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    put:
      operationId: renameAccountCompatibility
      deprecated: true
      summary: Alias temporal de compatibilidad para renombrar una cuenta
      description: Usa PATCH como método canónico. PUT se mantiene temporalmente para clientes antiguos.
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [display_name]
              properties:
                display_name: { type: string, minLength: 2, maxLength: 48 }
      responses:
        '200':
          description: Cuenta actualizada
          content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } }
        '400':
          description: Nombre inválido
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Sesión inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Cuenta inexistente o ajena
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/accounts/{account_id}/balance:
    get:
      operationId: getAccountBalance
      summary: Lee el saldo desde TigerBeetle
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountID'
      responses:
        '200':
          description: Saldos en unidades menores
          content: { application/json: { schema: { $ref: '#/components/schemas/Balance' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Cuenta inexistente o ajena
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/accounts/{account_id}/transactions:
    get:
      operationId: listAccountTransactions
      summary: Lista el historial reciente de una cuenta
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountID'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: cursor
          in: query
          description: Cursor opaco devuelto por la página anterior; no debe interpretarse ni modificarse
          schema: { type: string }
      responses:
        '200':
          description: Transferencias más recientes primero
          content: { application/json: { schema: { $ref: '#/components/schemas/HistoryResponse' } } }
        '400':
          description: Límite o cursor inválido
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '403':
          description: MFA no completado o cuenta ajena
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/accounts/{account_id}/transactions.csv:
    get:
      operationId: exportAccountTransactionsCsv
      summary: Exporta el historial reciente en CSV
      description: Exporta como máximo 100 movimientos después de comprobar la propiedad de la cuenta.
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountID'
      responses:
        '200':
          description: Archivo CSV de movimientos
          content:
            text/csv:
              schema: { type: string, format: binary }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Cuenta inexistente o ajena
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/accounts/{account_id}/deposits:
    post:
      operationId: createDeposit
      summary: Acredita fondos mediante la cuenta de compensación
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MovementRequest' } } }
      responses:
        '200':
          description: Operación registrada
          content: { application/json: { schema: { $ref: '#/components/schemas/Operation' } } }
        '400':
          description: Solicitud o clave de idempotencia inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '409':
          description: Clave reutilizada con otra solicitud
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '403':
          description: Canal de depósito demo deshabilitado
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Cuenta inexistente, inactiva o no autorizada
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '503':
          description: Ledger no disponible
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/accounts/{account_id}/withdrawals:
    post:
      operationId: createWithdrawal
      summary: Debita fondos hacia la cuenta de compensación
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AccountID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MovementRequest' } } }
      responses:
        '200':
          description: Operación registrada
          content: { application/json: { schema: { $ref: '#/components/schemas/Operation' } } }
        '422':
          description: Fondos insuficientes
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Cuenta inexistente, inactiva o no autorizada; el código público es account_not_found
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '403':
          description: MFA no completado o ruta financiera no autorizada
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '409':
          description: Clave de idempotencia reutilizada con otra solicitud
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '503':
          description: Ledger no disponible
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/transfers:
    post:
      operationId: createTransfer
      summary: Transfiere fondos entre cuentas propias o externas USD
      description: transfer_type=own exige ownership en origen y destino; transfer_type=external exige una cuenta destino activa de otro usuario. Ambos flujos requieren Idempotency-Key.
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TransferRequest' } } }
      responses:
        '200':
          description: Transferencia registrada
          content: { application/json: { schema: { $ref: '#/components/schemas/Operation' } } }
        '422':
          description: Fondos insuficientes o cuenta inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Cuenta origen o destino inexistente, inactiva o no autorizada
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '409':
          description: Una transferencia externa requiere un PIN de confirmación vigente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '400':
          description: Solicitud, moneda, tipo de transferencia o clave de idempotencia inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '403':
          description: MFA no completado o PIN de transferencia inválido
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '429':
          description: PIN temporalmente bloqueado por demasiados intentos
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '503':
          description: Ledger o almacén de PIN no disponible
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    AccountID:
      name: account_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Mensaje público y no sensible del error
          example: invalid credentials
        code:
          type: string
          description: Código estable para clientes
          example: insufficient_funds
...
    CreateAccountRequest:
      type: object
      additionalProperties: false
      properties:
        currency:
          type: string
          enum: [USD]
          default: USD
    MovementRequest:
      type: object
      additionalProperties: false
      required: [amount]
      properties:
        amount:
          type: string
          pattern: '^[1-9][0-9]*$'
          description: Unidades menores enteras, nunca un número de punto flotante
          example: '125000'
        currency:
          type: string
          enum: [USD]
          default: USD
    TransferRequest:
      allOf:
        - $ref: '#/components/schemas/MovementRequest'
        - type: object
          required: [source_account_id, destination_account_id]
          properties:
            source_account_id: { type: string, format: uuid }
            destination_account_id: { type: string, format: uuid }
            transfer_type:
              type: string
              enum: [own, external]
              default: own
              description: own exige ownership en ambas cuentas; external exige una cuenta activa de otro usuario.
            confirmation_pin:
              type: string
              pattern: '^[0-9]{4}$'
              writeOnly: true
              description: PIN requerido únicamente para transferencias externas.
    Account:
      type: object
      required: [id, display_name, currency, type, status, created_at]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string, maxLength: 48 }
        currency: { type: string, enum: [USD] }
        type: { type: string, enum: [checking] }
        status: { type: string, enum: [provisioning, active, failed, closed] }
        created_at: { type: string, format: date-time }
    Balance:
      type: object
      required: [account_id, currency, balance, available_balance, credits_posted, debits_posted, credits_pending, debits_pending]
      properties:
        account_id: { type: string, format: uuid }
        currency: { type: string, enum: [USD] }
        balance: { type: string, pattern: '^[0-9]+$' }
        available_balance: { type: string, pattern: '^[0-9]+$' }
        credits_posted: { type: string, pattern: '^[0-9]+$' }
        debits_posted: { type: string, pattern: '^[0-9]+$' }
        credits_pending: { type: string, pattern: '^[0-9]+$' }
        debits_pending: { type: string, pattern: '^[0-9]+$' }
    Operation:
      type: object
      required: [id, type, status, transfer_id, amount, currency, created_at]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [deposit, withdrawal, transfer] }
        status: { type: string, enum: [succeeded] }
        transfer_id: { type: string }
        amount: { type: string, pattern: '^[1-9][0-9]*$' }
        currency: { type: string, enum: [USD] }
        created_at: { type: string, format: date-time }
    Transaction:
      type: object
      required: [transfer_id, type, direction, amount, currency, created_at]
      properties:
        transfer_id: { type: string }
        type: { type: string, enum: [deposit, withdrawal, transfer] }
        direction: { type: string, enum: [credit, debit] }
        amount: { type: string, pattern: '^[1-9][0-9]*$' }
        currency: { type: string, enum: [USD] }
        created_at: { type: string, format: date-time }
    HistoryResponse:
      type: object
      required: [items, has_more]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }
        has_more: { type: boolean }
        next_cursor: { type: string, description: Cursor opaco para solicitar la siguiente página }
...